Fixed formatting issue

+++ Classes/RootViewController.m
@@ -75,7 +75,7 @@
     }
     
     // Configure the cell.
-    cell.textLabel.text = [NSString stringWithFormat:@"Row %d", indexPath.row];
+    cell.textLabel.text = [NSString stringWithFormat:@"Row %ld", (long)indexPath.row];
     return cell;
 }
 
@@ -87,7 +87,7 @@
 - (void)tableView:(UITableView *)aTableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
 {
 	// When a row is selected, set the detail view controller's detail item to the item associated with the selected row.
-    detailViewController.detailItem = [NSString stringWithFormat:@"Row %d", indexPath.row];
+    detailViewController.detailItem = [NSString stringWithFormat:@"Row %ld", (long)indexPath.row];
 }
